Refactor google credentials into a shared function (#4893)
